Add files via upload
6/30

1. 이제 AI가 죽을 때를 구현하여 보자 -> 일단 AI가 총에 맞으면 (현재 한대맞으면 죽게함) 죽는거랑 AI를 멈추는 거 구현완료
2. 근데 현재 지금 플레이어가 공격 키를 눌러도 공격이 간헐적으로 안나간다(좀 심하게 안나감)
3. 2번 조치한후에 AI 몬스터의 Hp에 맞춰서 바로 안죽게 하고 스테이지를 구현해보자

+++ Source/ZombieShooting/MyGameInstance.cpp
@@ -118,10 +118,10 @@ FString UMyGameInstance::GetMonsterName(FString MonsterType)
 	return MonsterName;
 }
 
-int UMyGameInstance::GetMonsterHp(FString MonsterType)
+float UMyGameInstance::GetMonsterHp(FString MonsterType)
 {
 	FMonsterDataTable* MonsterData = FMonsterTable->FindRow<FMonsterDataTable>(*MonsterType, TEXT(""));
-	int MonsterHp = MonsterData->MonsterHp;
+	float MonsterHp = MonsterData->MonsterHp;
 	return MonsterHp;
 }
